Update BooksContainer.js file
